added authorization for blogs

+++ blogapp/views.py
@@ -1,7 +1,9 @@
 from django.contrib.auth.models import User
+from django.core.exceptions import PermissionDenied
 from django.shortcuts import get_object_or_404, redirect, render
 from django.http import HttpResponse
 from django.views import generic
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
 from django.contrib.auth import login, logout
 
@@ -11,6 +13,7 @@ from .models import Blog
 
 class IndexView(generic.ListView):
     model = Blog
+    queryset = Blog.objects.select_related('author')
     template_name = 'blogapp/index.html'
     context_object_name = 'blogs'
 
@@ -33,6 +36,7 @@ def login_view(request):
         form = AuthenticationForm()
     return render(request, 'blogapp/login.html', {'form': form})
 
+@login_required
 def logout_view(request):
     logout(request)
     return redirect('blogapp:index')
@@ -47,6 +51,7 @@ def register_view(request):
         form = UserCreationForm()
     return render(request, 'blogapp/register.html', {'form': form})
 
+@login_required
 def create_blog(request):
     if request.method == 'POST':
         form = BlogForm(request.POST)
@@ -59,8 +64,13 @@ def create_blog(request):
         form = BlogForm()
     return render(request, 'blogapp/create.html', {'form': form})
 
+@login_required
 def edit_blog(request, pk):
     blog = get_object_or_404(Blog, pk=pk)
+
+    if blog.author != request.user:
+        raise PermissionDenied
+
     if request.method == 'POST':
         form = BlogForm(request.POST, instance=blog)
         if form.is_valid():
@@ -70,7 +80,12 @@ def edit_blog(request, pk):
         form = BlogForm(instance=blog)
     return render(request, 'blogapp/edit.html', {'form': form, 'blog': blog})
 
+@login_required
 def delete_blog(request, pk):
     blog = get_object_or_404(Blog, pk=pk)
+
+    if blog.author != request.user:
+        raise PermissionDenied
+
     blog.delete()
     return redirect('blogapp:index')
